+++ HISTORY.md
@@ -4,6 +4,7 @@ History of features added to LiChess Tools in time (reversed order)
 
 - removed the "No study chat" theme (use "Toggle study chat")
 - added Ctrl-C as general key shortcut to copy the FEN
+- highlight Explorer moves that are in the move list as next
 
 19 Jun